Add tile group with example, rename `setScollBar` to `setScrollbar`
This PR also includes the following other changes:
 * Removal of old comments
 * Formatting with zig fmt
 * Refactoring of a couple lines to make them more readable
 * Addition of `Group.setScollbarSize` method
 * Fix a couple types

The thinking behing allowing signed X and Y values allows widgets to be
pushed off-screen, which makes sense for animations and whatnot.
However, negative sizes are undefined, so negative W and H should never
be allowed. My guess as to why FLTK uses `int` for all 4 values is
probably just because it's an easier type to work in C++ with than `unsigned
int` without requiring constant casting